% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_competition}
\alias{get_competition}
\title{Get competition Function}
\usage{
get_competition(competition_id)
}
\description{
This function allows you to get competition information for a given competition_id.
}
