% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_leagues}
\alias{get_leagues}
\title{Get All Leagues Function}
\usage{
get_leagues()
}
\description{
This function allows you to get all leagues you currently have access to.
}
\examples{
get_leagues()
}
\keyword{Leagues}
