% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_players_competition}
\alias{get_players_competition}
\title{Get Players Function}
\usage{
get_players_competition(competition_id)
}
\description{
This function allows you to get all players from a specific competition.
}
